+++ R/dimensional_reduction.R
@@ -65,6 +65,10 @@ JackStraw <- function(
     my.sapply <- future_sapply
   }
   assay <- assay %||% DefaultAssay(object = object)
+  if (IsSCT(assay = object[[assay]])) {
+    stop("JackStraw cannot be run on SCTransform-normalized data.
+         Please supply a non-SCT assay.")
+  }
   if (dims > length(x = object[[reduction]])) {
     dims <- length(x = object[[reduction]])
     warning("Number of dimensions specified is greater than those available. Setting dims to ", dims, " and continuing", immediate. = TRUE)
@@ -958,14 +962,10 @@ RunPCA.Seurat <- function(
 #' \item{FIt-SNE: }{Use the FFT-accelerated Interpolation-based t-SNE. Based on
 #' Kluger Lab code found here: https://github.com/KlugerLab/FIt-SNE}
 #' }
-#' @param add.iter If an existing tSNE has already been computed, uses the
-#' current tSNE to seed the algorithm and then adds additional iterations on top
-#' of this
 #' @param dim.embed The dimensional space of the resulting tSNE embedding
 #' (default is 2). For example, set to 3 for a 3d tSNE
 #' @param reduction.key dimensional reduction key, specifies the string before the number for the dimension names. tSNE_ by default
 #'
-#' @importFrom tsne tsne
 #' @importFrom Rtsne Rtsne
 #'
 #' @rdname RunTSNE
@@ -977,7 +977,6 @@ RunTSNE.matrix <- function(
   assay = NULL,
   seed.use = 1,
   tsne.method = "Rtsne",
-  add.iter = 0,
   dim.embed = 2,
   reduction.key = "tSNE_",
   ...
@@ -995,14 +994,6 @@ RunTSNE.matrix <- function(
     'FIt-SNE' = fftRtsne(X = object, dims = dim.embed, rand_seed = seed.use, ...),
     stop("Invalid tSNE method: please choose from 'Rtsne' or 'FIt-SNE'")
   )
-  if (add.iter > 0) {
-    tsne.data <- tsne(
-      X = object,
-      initial_config = as.matrix(x = tsne.data),
-      max_iter = add.iter,
-      ...
-    )
-  }
   colnames(x = tsne.data) <- paste0(reduction.key, 1:ncol(x = tsne.data))
   rownames(x = tsne.data) <- rownames(x = object)
   tsne.reduction <- CreateDimReducObject(
@@ -1027,7 +1018,6 @@ RunTSNE.DimReduc <- function(
   dims = 1:5,
   seed.use = 1,
   tsne.method = "Rtsne",
-  add.iter = 0,
   dim.embed = 2,
   reduction.key = "tSNE_",
   ...
@@ -1050,7 +1040,6 @@ RunTSNE.dist <- function(
   assay = NULL,
   seed.use = 1,
   tsne.method = "Rtsne",
-  add.iter = 0,
   dim.embed = 2,
   reduction.key = "tSNE_",
   ...
@@ -1083,7 +1072,6 @@ RunTSNE.Seurat <- function(
   features = NULL,
   seed.use = 1,
   tsne.method = "Rtsne",
-  add.iter = 0,
   dim.embed = 2,
   distance.matrix = NULL,
   reduction.name = "tsne",
@@ -1097,7 +1085,6 @@ RunTSNE.Seurat <- function(
       assay = DefaultAssay(object = object),
       seed.use = seed.use,
       tsne.method = tsne.method,
-      add.iter = add.iter,
       dim.embed = dim.embed,
       reduction.key = reduction.key,
       is_distance = TRUE,
@@ -1110,7 +1097,6 @@ RunTSNE.Seurat <- function(
       dims = dims,
       seed.use = seed.use,
       tsne.method = tsne.method,
-      add.iter = add.iter,
       dim.embed = dim.embed,
       reduction.key = reduction.key,
       pca = FALSE,
@@ -1122,7 +1108,6 @@ RunTSNE.Seurat <- function(
       assay = DefaultAssay(object = object),
       seed.use = seed.use,
       tsne.method = tsne.method,
-      add.iter = add.iter,
       dim.embed = dim.embed,
       reduction.key = reduction.key,
       pca = FALSE,
@@ -1137,7 +1122,7 @@ RunTSNE.Seurat <- function(
 }
 
 #' @importFrom reticulate py_module_available py_set_seed import
-#' @importFrom uwot umap
+#' @importFrom uwot umap umap_transform
 #' @importFrom future nbrOfWorkers
 #'
 #' @rdname RunUMAP
@@ -1146,6 +1131,7 @@ RunTSNE.Seurat <- function(
 #'
 RunUMAP.default <- function(
   object,
+  reduction.model = NULL,
   assay = NULL,
   umap.method = 'uwot',
   n.neighbors = 30L,
@@ -1245,8 +1231,75 @@ RunUMAP.default <- function(
         verbose = verbose
       )
     },
+    'uwot-learn' = {
+      if (metric == 'correlation') {
+        warning(
+          "UWOT does not implement the correlation metric, using cosine instead",
+          call. = FALSE,
+          immediate. = TRUE
+        )
+        metric <- 'cosine'
+      }
+      umap(
+        X = object,
+        n_threads = nbrOfWorkers(),
+        n_neighbors = as.integer(x = n.neighbors),
+        n_components = as.integer(x = n.components),
+        metric = metric,
+        n_epochs = n.epochs,
+        learning_rate = learning.rate,
+        min_dist = min.dist,
+        spread = spread,
+        set_op_mix_ratio = set.op.mix.ratio,
+        local_connectivity = local.connectivity,
+        repulsion_strength = repulsion.strength,
+        negative_sample_rate = negative.sample.rate,
+        a = a,
+        b = b,
+        fast_sgd = uwot.sgd,
+        verbose = verbose,
+        ret_model = TRUE
+      )
+    },
+    'uwot-predict' = {
+      if (metric == 'correlation') {
+        warning(
+          "UWOT does not implement the correlation metric, using cosine instead",
+          call. = FALSE,
+          immediate. = TRUE
+        )
+        metric <- 'cosine'
+      }
+      if (is.null(x = reduction.model) || !inherits(x = reduction.model, what = 'DimReduc')) {
+        stop(
+          "If using uwot-predict, please pass a DimReduc object with the model stored to reduction.model.",
+          call. = FALSE
+        )
+      }
+      model <- reduction.model %||% Misc(
+        object = reduction.model,
+        slot = "model"
+      )
+      if (length(x = model) == 0) {
+        stop(
+          "The provided reduction.model does not have a model stored. Please try running umot-learn on the object first",
+          call. = FALSE
+        )
+      }
+      umap_transform(
+        X = object,
+        model = model,
+        n_threads = nbrOfWorkers(),
+        n_epochs = n.epochs,
+        verbose = verbose
+      )
+    },
     stop("Unknown umap method: ", umap.method, call. = FALSE)
   )
+  if (umap.method == 'uwot-learn') {
+    umap.model <- umap.output
+    umap.output <- umap.output$embedding
+  }
   colnames(x = umap.output) <- paste0(reduction.key, 1:ncol(x = umap.output))
   if (inherits(x = object, what = 'dist')) {
     rownames(x = umap.output) <- attr(x = object, "Labels")
@@ -1259,6 +1312,9 @@ RunUMAP.default <- function(
     assay = assay,
     global = TRUE
   )
+  if (umap.method == 'uwot-learn') {
+    Misc(umap.reduction, slot = "model") <- umap.model
+  }
   return(umap.reduction)
 }
 
@@ -1350,6 +1406,7 @@ RunUMAP.Graph <- function(
   return(umap)
 }
 
+#' @param reduction.model \code{DimReduc} object that contains the umap model
 #' @param dims Which dimensions to use as input features, used only if
 #' \code{features} is NULL
 #' @param reduction Which dimensional reduction (PCA or ICA) to use for the
@@ -1363,6 +1420,7 @@ RunUMAP.Graph <- function(
 #' @param umap.method UMAP implementation to run. Can be
 #' \describe{
 #'   \item{\code{uwot}:}{Runs umap via the uwot R package}
+#'   \item{\code{uwot-learn}:}{Runs umap via the uwot R package and return the learned umap model}
 #'   \item{\code{umap-learn}:}{Run the Seurat wrapper of the python umap-learn package}
 #' }
 #' @param n.neighbors This determines the number of neighboring points used in
@@ -1423,6 +1481,7 @@ RunUMAP.Graph <- function(
 #'
 RunUMAP.Seurat <- function(
   object,
+  reduction.model = NULL,
   dims = NULL,
   reduction = 'pca',
   features = NULL,
@@ -1487,6 +1546,7 @@ RunUMAP.Seurat <- function(
   }
   object[[reduction.name]] <- RunUMAP(
     object = data.use,
+    reduction.model = reduction.model,
     assay = assay,
     umap.method = umap.method,
     n.neighbors = n.neighbors,