+++ R/objects.R
@@ -1320,6 +1320,12 @@ RenameAssays <- function(object, ...) {
       DefaultAssay(object = object) <- new
     }
     Key(object = object[[new]]) <- old.key
+    # change assay used in any dimreduc object
+    for (i in Reductions(object = object)) {
+      if (DefaultAssay(object = object[[i]]) == old) {
+        DefaultAssay(object = object[[i]]) <- new
+      }
+    }
     object[[old]] <- NULL
   }
   return(object)
@@ -3360,6 +3366,18 @@ Misc.Assay <- function(object, slot = NULL, ...) {
 
 #' @rdname Misc
 #' @export
+#' @method Misc DimReduc
+#'
+Misc.DimReduc <- function(object, slot = NULL, ...) {
+  CheckDots(...)
+  if (is.null(x = slot)) {
+    return(slot(object = object, name = 'misc'))
+  }
+  return(slot(object = object, name = 'misc')[[slot]])
+}
+
+#' @rdname Misc
+#' @export
 #' @method Misc Seurat
 #'
 #' @examples
@@ -3380,6 +3398,23 @@ Misc.Seurat <- function(object, slot = NULL, ...) {
 #' @method Misc<- Assay
 #'
 "Misc<-.Assay" <- function(object, slot, ..., value) {
+  CheckDots(...)
+  if (slot %in% names(x = Misc(object = object))) {
+    warning("Overwriting miscellanous data for ", slot)
+  }
+  if (is.list(x = value)) {
+    slot(object = object, name = 'misc')[[slot]] <- c(value)
+  } else {
+    slot(object = object, name = 'misc')[[slot]] <- value
+  }
+  return(object)
+}
+
+#' @rdname Misc
+#' @export
+#' @method Misc<- DimReduc
+#'
+"Misc<-.DimReduc" <- function(object, slot, ..., value) {
   CheckDots(...)
   if (slot %in% names(x = Misc(object = object))) {
     warning("Overwriting miscellanous data for ", slot)
@@ -4803,6 +4838,7 @@ VariableFeatures.Seurat <- function(object, assay = NULL, selection.method = NUL
 #' @param invert Invert the selection of cells
 #'
 #' @importFrom stats na.omit
+#' @importFrom rlang is_quosure enquo eval_tidy
 #'
 #' @rdname WhichCells
 #' @export
@@ -4819,12 +4855,14 @@ WhichCells.Assay <- function(
   cells <- cells %||% colnames(x = object)
   if (!missing(x = expression) && !is.null(x = substitute(expr = expression))) {
     key.pattern <- paste0('^', Key(object = object))
-    expr <- if (is.call(x = substitute(expr = expression))) {
-      substitute(expr = expression)
+    expr <- if (tryCatch(expr = is_quosure(x = expression), error = function(...) FALSE)) {
+      expression
+    } else if (is.call(x = enquo(arg = expression))) {
+      enquo(arg = expression)
     } else {
       parse(text = expression)
     }
-    expr.char <- as.character(x = expr)
+    expr.char <- suppressWarnings(expr = as.character(x = expr))
     expr.char <- unlist(x = lapply(X = expr.char, FUN = strsplit, split = ' '))
     expr.char <- gsub(
       pattern = key.pattern,
@@ -4847,8 +4885,7 @@ WhichCells.Assay <- function(
     expr.char <- expr.char[vars.use]
     data.subset <- as.data.frame(x = t(x = as.matrix(x = object[expr.char, ])))
     colnames(x = data.subset) <- expr.char
-    data.subset <- subset.data.frame(x = data.subset, subset = eval(expr = expr))
-    cells <- rownames(x = data.subset)
+    cells <- rownames(x = data.subset)[eval_tidy(expr = expr, data = data.subset)]
   }
   if (invert) {
     cells <- colnames(x = object)[!colnames(x = object) %in% cells]
@@ -4865,6 +4902,7 @@ WhichCells.Assay <- function(
 #' @param seed Random seed for downsampling. If NULL, does not set a seed
 #'
 #' @importFrom stats na.omit
+#' @importFrom rlang is_quosure enquo eval_tidy
 #'
 #' @rdname WhichCells
 #' @export
@@ -4923,12 +4961,14 @@ WhichCells.Seurat <- function(
       }
     )
     key.pattern <- paste0('^', object.keys, collapse = '|')
-    expr <- if (is.call(x = substitute(expr = expression))) {
-      substitute(expr = expression)
+    expr <- if (tryCatch(expr = is_quosure(x = expression), error = function(...) FALSE)) {
+      expression
+    } else if (is.call(x = enquo(arg = expression))) {
+      enquo(arg = expression)
     } else {
       parse(text = expression)
     }
-    expr.char <- as.character(x = expr)
+    expr.char <- suppressWarnings(expr = as.character(x = expr))
     expr.char <- unlist(x = lapply(X = expr.char, FUN = strsplit, split = ' '))
     expr.char <- gsub(
       pattern = '(',
@@ -4948,12 +4988,11 @@ WhichCells.Seurat <- function(
     )
     data.subset <- FetchData(
       object = object,
-      vars = expr.char[vars.use],
+      vars = unique(x = expr.char[vars.use]),
       cells = cells,
       slot = slot
     )
-    data.subset <- subset.data.frame(x = data.subset, subset = eval(expr = expr))
-    cells <- rownames(x = data.subset)
+    cells <- rownames(x = data.subset)[eval_tidy(expr = expr, data = data.subset)]
   }
   if (invert) {
     cell.order <- colnames(x = object)
@@ -6089,7 +6128,7 @@ subset.Assay <- function(x, cells = NULL, features = NULL, ...) {
     replacement = '',
     x = features
   )
-  features <- intersect(x = rownames(x = x), y = features)
+  features <- intersect(x = features, y = rownames(x = x))
   if (length(x = features) == 0) {
     stop("Cannot find features provided")
   }
@@ -6109,6 +6148,30 @@ subset.Assay <- function(x, cells = NULL, features = NULL, ...) {
   }
   VariableFeatures(object = x) <- VariableFeatures(object = x)[VariableFeatures(object = x) %in% features]
   slot(object = x, name = 'meta.features') <- x[[]][features, , drop = FALSE]
+  if (IsSCT(assay = x)) {
+    # subset cells and genes in the SCT assay
+    obj.misc <- Misc(object = x)
+    if ("vst.set" %in% names(x = obj.misc)) {
+      # set of vst.out objects
+      vst.info <- obj.misc[["vst.set"]]
+      for (i in seq_along(along.with = vst.info)) {
+        vst.info[[i]] <- SubsetVST(
+          sct.info = vst.info[[i]],
+          cells = cells,
+          features = features
+        )
+      }
+      obj.misc[["vst.set"]] <- vst.info
+    } else {
+      # just one vst.out
+      obj.misc[["vst.out"]] <- SubsetVST(
+        sct.info = obj.misc[["vst.out"]],
+        cells = cells,
+        features = features
+      )
+    }
+    slot(object = x, name = "misc") <- obj.misc
+  }
   return(x)
 }
 
@@ -6184,6 +6247,8 @@ subset.DimReduc <- function(x, cells = NULL, features = NULL, ...) {
 #'
 #' @return A subsetted Seurat object
 #'
+#' @importFrom rlang enquo
+#'
 #' @rdname subset.Seurat
 #' @aliases subset
 #' @seealso \code{\link[base]{subset}} \code{\link{WhichCells}}
@@ -6201,7 +6266,7 @@ subset.DimReduc <- function(x, cells = NULL, features = NULL, ...) {
 subset.Seurat <- function(x, subset, cells = NULL, features = NULL, idents = NULL, ...) {
   x <- UpdateSlots(object = x)
   if (!missing(x = subset)) {
-    subset <- deparse(expr = substitute(expr = subset))
+    subset <- enquo(arg = subset)
   }
   cells <- WhichCells(
     object = x,
@@ -6999,6 +7064,19 @@ Projected <- function(object) {
   return(!all(projected.dims == 0))
 }
 
+# Subset cells in vst data
+# @param sct.info A vst.out list 
+# @param cells vector of cells to retain
+# @param features vector of features to retain
+SubsetVST <- function(sct.info, cells, features) {
+  cells.keep <- intersect(x = cells, y = rownames(x = sct.info$cell_attr))
+  sct.info$cell_attr <- sct.info$cell_attr[cells.keep, ]
+  # find which subset of features are in the SCT assay
+  feat.keep <- intersect(x = features, y = rownames(x = sct.info$gene_attr))
+  sct.info$gene_attr <- sct.info$gene_attr[feat.keep, ]
+  return(sct.info)
+}
+
 # Get the top
 #
 # @param data Data to pull the top from
@@ -7216,6 +7294,7 @@ ValidateDataForMerge <- function(assay, slot) {
       ncol = data.dims[2],
       dimnames = dimnames(x = GetAssayData(object = assay, slot = data.slot))
     )
+    mat <- as(object = mat, Class = "dgCMatrix")
   }
   return(mat)
 }