+++ tests/testthat/test_differential_expression.R
@@ -168,7 +168,8 @@ test_that("bimod test works", {
 results <- FindMarkers(object = pbmc_small, ident.1 = 0, ident.2 = 1, test.use = "roc", verbose = FALSE)
 test_that("roc test works", {
   expect_equal(nrow(x = results), 201)
-  expect_equal(colnames(x = results), c("myAUC", "avg_diff", "power", "pct.1", "pct.2"))
+  # expect_equal(colnames(x = results), c("myAUC", "avg_diff", "power", "pct.1", "pct.2"))
+  expect_equal(colnames(x = results), c("myAUC", "avg_diff", "power", "avg_logFC", "pct.1", "pct.2"))
   expect_equal(results["CST3", "myAUC"], 0.018)
   expect_equal(results["CST3", "avg_diff"], -2.552769, tolerance = 1e-6)
   expect_equal(results["CST3", "power"], 0.964)